Proposal: Pre-Build images in Github
This is a proposal as requested by @yanokwa at
https://forum.getodk.org/t/host-odk-central-docker-images-on-github-container-registry/29812/19?u=mattelacchiato

It introduces the usage of Github actions to build the docker images in
Github and host them at ghcr.io (the github package repository). For
open source projects, the infrastructure is for free.

Additionally, I've removed the unnamed volumes from the
docker-compose.yml. The reason is that unnamed volumes are hard to
backup. Currently, the enketo secrets are not part of the backup, which
is really bad if you have to recover your installation and all public
links are not working anymore for submitters.

For the same aspect of easier backup on OS-level, I moved the postgres
and redis folder to a local volume, too.

I switch the secrets docker base image from node to alpine, which is
much more lightweight. We don't need node for a simple bash script
execution.

This setup will increase the benefits for experienced administrators
without removing any convenience for beginners.