service/start: check check-migrations failure more robustly
If check-migrations fails with non-1 exit code, the server will continue
to start.

This check was originally introduced as a fix for #461.  This change should
be more robust than the existing approach, but `set -e` and `set -o pipefail`
might provide even more safety in handling failures in this script.

There may also be other critical scripts in this repo which would benefit
from similar changes.