make dev: ensure running code is up-to-date (#1422)
Currently, `make dev` will not reflect local changes to docker config, files etc.

# This change:

The official `docker compose` docs ref the `--build` flag are scant: https://docs.docker.com/reference/cli/docker/compose/up/

However, behaviour seems to be that if `--build` is supplied, then containers with a `build:` section defined in any of the compose `yaml` files will be rebuilt.

This has the effect that changes in any of the following could trigger a rebuild:

* `files/`
* the relevant `.dockerfile`
* the compose `.yaml` file(s)

If no changes are detected, this "rebuild" seems to be just a fetch from the local cache.

This commit also expands the `-d` flat to long-form `--detach`.